装备系统

+++ Assets/Scripts/UIRelated/HandScript.cs
@@ -39,7 +39,11 @@ public class HandScript : MonoBehaviour
     {
         icon.transform.position = Input.mousePosition + offset;
 
-        DeleteItem();
+        if (Input.GetMouseButton(0) && !EventSystem.current.IsPointerOverGameObject() && MyInstance.MyMoveable != null)
+        {
+            Debug.Log("hhhh");
+            DeleteItem();
+        }
     }
 
     public void TakeMoveable(IMoveable moveable)
@@ -65,18 +69,15 @@ public class HandScript : MonoBehaviour
         icon.color = new Color(0, 0, 0, 0);
     }
 
-    private void DeleteItem()
+    public void DeleteItem()
     {
-        if (Input.GetMouseButton(0) && !EventSystem.current.IsPointerOverGameObject() && MyInstance.MyMoveable != null)
+        if (MyMoveable is Item && InventotyScript.MyInstance.FromSlot != null)
         {
-            if (MyMoveable is Item && InventotyScript.MyInstance.FromSlot != null) 
-            {
-                (MyMoveable as Item).MySlot.Clear();
-            }
-
-            Drop();
-
-            InventotyScript.MyInstance.FromSlot = null;
+            (MyMoveable as Item).MySlot.Clear();
         }
+
+        Drop();
+
+        InventotyScript.MyInstance.FromSlot = null;
     }
 }
